requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
requests==2.25.1
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.5
vk-api==11.8.0
Werkzeug==1.0.1
numpy~=1.19.1